+++ org.sonar.ide.eclipse.erlide/META-INF/MANIFEST.MF
@@ -5,7 +5,7 @@ Bundle-SymbolicName: org.sonar.ide.eclipse.erlide;singleton:=true
 Bundle-Version: 3.0.0.qualifier
 Bundle-Vendor: %bundle_vendor
 Bundle-ClassPath: .
-Bundle-Activator: org.sonar.ide.eclipse.erlide.internal.SonarPyDevPlugin
+Bundle-Activator: org.sonar.ide.eclipse.erlide.internal.SonarErlIdePlugin
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.core.resources,
  org.eclipse.ui,
@@ -15,8 +15,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.sonar.ide.eclipse.ui,
  org.sonar.ide.eclipse.core,
  org.sonar.ide.eclipse.wsclient,
- org.python.pydev,
- org.python.pydev.core
+ org.erlide.core
 Bundle-RequiredExecutionEnvironment: J2SE-1.5
 Bundle-ActivationPolicy: lazy
 Bundle-Localization: OSGI-INF/l10n/bundle
